refactor: 패키지 생성
domain 로직 과 UI 분리
