+++ src/main/java/baseball/Application.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import baseball.ui.PlayGrounds;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
