+++ src/main/java/baseball/domain/Ball.java
@@ -1,6 +1,4 @@
-package baseball;
-
-import java.util.Objects;
+package baseball.domain;
 
 public class Ball {
