+++ src/main/java/baseball/domain/BallStatus.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 public enum BallStatus {
     STRIKE, BALL, NOTHING;
