+++ src/main/java/baseball/domain/InningStatus.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 public class InningStatus {
