+++ src/main/java/baseball/domain/Pitcher.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 import static camp.nextstep.edu.missionutils.Randoms.pickNumberInRange;
