+++ src/test/java/baseball/domain/PitcherTest.java
@@ -1,7 +1,8 @@
-package baseball;
+package baseball.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.domain.Pitcher;
 import java.util.HashSet;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
